+++ changelog.md
@@ -17,6 +17,7 @@ Codion Change Log
 ### is.codion.swing.common.ui
 - FilterTableCellRenderer.Builder.ColumnClassBuilder added.
 - FilterTable.Builder.rowHeight(), rowMargin(), intercellSpacing(), gridColor(), showGrid(), showHorizontalLines(), showVerticalLines() and dragEnabled() added.
+- ConditionPanel.focusGainedObserver() renamed focusGained().
 ### is.codion.framework.db.local
 - MethodLogger renamed MethodTracer, logger package renamed tracer.
 - DefaultLocalEntityConnection.NoOpTracer added, tracer no longer nullable.